ref(node): Extract cacheResponseHook to a module free of OTel redis instrumentation
The orchestrion opt-in imports cacheResponseHook; importing it from redis/index
transitively pulled the vendored OTel IORedisInstrumentation/RedisInstrumentation
into the opt-in module graph. Move cacheResponseHook + _redisOptions into
redis/cache.ts (no OTel instrumentation imports) so apps using injection only for
mysql/lru-memoizer don't bundle the redis OTel instrumentation.
