+++ packages/node/src/integrations/tracing/redis/index.ts
@@ -1,107 +1,19 @@
-import type { IntegrationFn, Span } from '@sentry/core';
-import {
-  defineIntegration,
-  SEMANTIC_ATTRIBUTE_CACHE_HIT,
-  SEMANTIC_ATTRIBUTE_CACHE_ITEM_SIZE,
-  SEMANTIC_ATTRIBUTE_CACHE_KEY,
-  SEMANTIC_ATTRIBUTE_SENTRY_OP,
-  spanToJSON,
-  truncate,
-  waitForTracingChannelBinding,
-} from '@sentry/core';
+import type { IntegrationFn } from '@sentry/core';
+import { defineIntegration, waitForTracingChannelBinding } from '@sentry/core';
 import * as dc from 'node:diagnostics_channel';
 import { subscribeRedisDiagnosticChannels, type RedisTracingChannelFactory } from '@sentry/server-utils';
 import { generateInstrumentOnce } from '@sentry/node-core';
 import { isDiagnosticsChannelInjectionEnabled } from '../../../sdk/diagnosticsChannelInjection';
-import type { IORedisCommandArgs } from '../../../utils/redisCache';
-import {
-  calculateCacheItemSize,
-  GET_COMMANDS,
-  getCacheKeySafely,
-  getCacheOperation,
-  isInCommands,
-  shouldConsiderForCache,
-} from '../../../utils/redisCache';
-import type { IORedisResponseCustomAttributeFunction } from './vendored/types';
+import { cacheResponseHook, type RedisOptions, setRedisOptions } from './cache';
 import { IORedisInstrumentation } from './vendored/ioredis-instrumentation';
 import { RedisInstrumentation } from './vendored/redis-instrumentation';
 
-interface RedisOptions {
-  /**
-   * Define cache prefixes for cache keys that should be captured as a cache span.
-   *
-   * Setting this to, for example, `['user:']` will capture cache keys that start with `user:`.
-   */
-  cachePrefixes?: string[];
-  /**
-   * Maximum length of the cache key added to the span description. If the key exceeds this length, it will be truncated.
-   *
-   * Passing `0` will use the full cache key without truncation.
-   *
-   * By default, the full cache key is used.
-   */
-  maxCacheKeyLength?: number;
-}
+// `cacheResponseHook`/`_redisOptions` live in `./cache` (which has no OTel
+// instrumentation imports) so the orchestrion opt-in can pull the hook without
+// dragging the OTel redis instrumentation in. Re-exported here for tests.
+export { _redisOptions, cacheResponseHook } from './cache';
 
 const INTEGRATION_NAME = 'Redis' as const;
-
-/* Only exported for testing purposes */
-export let _redisOptions: RedisOptions = {};
-
-/* Only exported for testing purposes */
-export const cacheResponseHook: IORedisResponseCustomAttributeFunction = (
-  span: Span,
-  redisCommand: string,
-  cmdArgs: IORedisCommandArgs,
-  response: unknown,
-) => {
-  const safeKey = getCacheKeySafely(redisCommand, cmdArgs);
-  const cacheOperation = getCacheOperation(redisCommand);
-
-  if (
-    !safeKey ||
-    !cacheOperation ||
-    !_redisOptions.cachePrefixes ||
-    !shouldConsiderForCache(redisCommand, safeKey, _redisOptions.cachePrefixes)
-  ) {
-    // not relevant for cache
-    return;
-  }
-
-  // otel/ioredis seems to be using the old standard, as there was a change to those params: https://github.com/open-telemetry/opentelemetry-specification/issues/3199
-  // We are using params based on the docs: https://opentelemetry.io/docs/specs/semconv/attributes-registry/network/
-  // Fall back to stable semconv attributes (server.address/server.port) when
-  // old-semconv ones are absent, eg OTEL_SEMCONV_STABILITY_OPT_IN=database
-  // set for node-redis v4/v5.
-  const spanData = spanToJSON(span).data;
-  const networkPeerAddress = spanData['net.peer.name'] ?? spanData['server.address'];
-  const networkPeerPort = spanData['net.peer.port'] ?? spanData['server.port'];
-  if (networkPeerPort && networkPeerAddress) {
-    span.setAttributes({ 'network.peer.address': networkPeerAddress, 'network.peer.port': networkPeerPort });
-  }
-
-  const cacheItemSize = calculateCacheItemSize(response);
-
-  if (cacheItemSize) {
-    span.setAttribute(SEMANTIC_ATTRIBUTE_CACHE_ITEM_SIZE, cacheItemSize);
-  }
-
-  if (isInCommands(GET_COMMANDS, redisCommand) && cacheItemSize !== undefined) {
-    span.setAttribute(SEMANTIC_ATTRIBUTE_CACHE_HIT, cacheItemSize > 0);
-  }
-
-  span.setAttributes({
-    [SEMANTIC_ATTRIBUTE_SENTRY_OP]: cacheOperation,
-    [SEMANTIC_ATTRIBUTE_CACHE_KEY]: safeKey,
-  });
-
-  // todo: change to string[] once EAP supports it
-  const spanDescription = safeKey.join(', ');
-
-  span.updateName(
-    _redisOptions.maxCacheKeyLength ? truncate(spanDescription, _redisOptions.maxCacheKeyLength) : spanDescription,
-  );
-};
 
 const instrumentIORedis = generateInstrumentOnce(`${INTEGRATION_NAME}.IORedis`, () => {
   return new IORedisInstrumentation({
@@ -151,7 +63,7 @@ const _redisIntegration = ((options: RedisOptions = {}) => {
   return {
     name: INTEGRATION_NAME,
     setupOnce() {
-      _redisOptions = options;
+      setRedisOptions(options);
       instrumentRedis();
     },
   };
